refactor(progress): purge formatter shims and micro-files
Closes #8824
