--- packages/extension/src/progressView/frontend/formatters/logFormatters/index.ts
@@ -1,37 +1,0 @@
-/**
- * Log formatters barrel export.
- *
- * Provides specialized formatters for different log message types.
- */
-
-// Banner formatters (thinking, scratchpad, model response)
-export {
-  formatBannerContentTemplate,
-  formatModelResponseTemplate,
-} from './bannerFormatters';
-
-// Tool formatters (tool use, web search, web fetch)
-export {
-  formatToolUseTemplate,
-  formatWebSearchTemplate,
-  formatWebFetchTemplate,
-} from './toolFormatters';
-
-// Data formatters (file list, missing outputs, latexdiff, statistics)
-export {
-  formatFileListTemplate,
-  formatMissingOutputsTemplate,
-  formatLatexdiffTemplate,
-  formatStatisticsTemplate,
-} from './dataFormatters';
-
-// Context management formatter
-export { formatContextManagementTemplate } from './contextManagementFormatters';
-
-// Message formatters (user message, progress, error, default)
-export {
-  formatUserMessageTemplate,
-  formatProgressStatusTemplate,
-  formatErrorTemplate,
-  formatDefaultLogMessageTemplate,
-} from './messageFormatters';
